Create service functions for deleting and patching log entries
